"Add buttons action and other fixes"
