+++ DeUrgenta/Features/CreateGroup/Coordinator/CreateGroupCoordinator.swift
@@ -1,0 +1,28 @@
+import Foundation
+import SwiftUI
+import UIKit
+
+final class CreateGroupCoordinator: Coordinator {
+    var navigationController: UINavigationController
+    var childCoordinators: [Coordinator] = []
+
+    init(navigationController: UINavigationController) {
+        self.navigationController = navigationController
+    }
+
+    func start() {
+        let viewController = UIHostingController(rootView: CreateGroupView(delegate: self))
+        navigationController.pushViewController(viewController, animated: true)
+    }
+}
+
+extension CreateGroupCoordinator: CreateGroupViewDelegate {
+    func CreateGroupViewDidTapAddGroup(_: CreateGroupView) {
+        let viewController = UIHostingController(rootView: EmptyView())
+        navigationController.pushViewController(viewController, animated: true)
+    }
+
+    func CreateGroupViewDidTapNoGroup(_: CreateGroupView) {
+        navigationController.popViewController(animated: true)
+    }
+}
