+++ DeUrgenta/Features/CreateGroup/View/CreateGroupView.swift
@@ -1,6 +1,13 @@
 import SwiftUI
 
+protocol CreateGroupViewDelegate {
+    func CreateGroupViewDidTapAddGroup(_ view: CreateGroupView)
+    func CreateGroupViewDidTapNoGroup(_ view: CreateGroupView)
+}
+
 struct CreateGroupView: View {
+    var delegate: CreateGroupViewDelegate
+
     var body: some View {
         VStack(alignment: .leading) {
             HStack {
@@ -8,12 +15,15 @@ struct CreateGroupView: View {
                     .font(.custom("IBMPlexSans-SemiBold", size: 24))
                     .padding(.leading, 20)
                     .padding(.top, 10)
+
                 Spacer()
             }
+
             VStack {
                 Text(AppStrings.CreateGroup.firstParagraph.localized())
                     .font(.custom("IBMPlexSans-Regular", size: 16))
                     .foregroundColor(.darkText)
+
                 Text(AppStrings.CreateGroup.secondParagraph.localized())
                     .padding(.top, 15)
                     .font(.custom("IBMPlexSans-Regular", size: 16))
@@ -23,10 +33,12 @@ struct CreateGroupView: View {
             .padding(.trailing, 20)
             .padding(.top, 5)
             .lineSpacing(10)
+
             Spacer()
+
             VStack {
                 Button(action: {
-                    // TODO: Implement this action
+                    delegate.CreateGroupViewDidTapNoGroup(self)
                 }, label: {
                     Text(AppStrings.CreateGroup.noGroup.localized())
                         .foregroundColor(.gray)
@@ -34,7 +46,7 @@ struct CreateGroupView: View {
                 })
 
                 Button(action: {
-                    // TODO: Implement this action
+                    delegate.CreateGroupViewDidTapAddGroup(self)
                 }, label: {
                     Text(AppStrings.CreateGroup.addGroup.localized())
                         .font(.custom("IBMPlexSans-Bold", size: 16))
@@ -49,11 +61,5 @@ struct CreateGroupView: View {
             .padding(.leading, 20)
             .padding(.trailing, 20)
         }
-    }
-}
-
-struct CreateGroupView_Previews: PreviewProvider {
-    static var previews: some View {
-        CreateGroupView()
     }
 }
