+++ DeUrgenta/Localization/Localizable.strings
@@ -12,7 +12,7 @@
 
 "CreateGroup.headerTitle" = "Creează-ți un grup";
 
-"CreateGroup.firstParagraph" = "În eventualitatea unui cutremur este important sa ai un protocol împreună cu cei apropiați prin care vă puteți regăsi cu ușurință. Creează un grup de prieteni, de familie sau de colegi de la birou sau școală pentru a vă seta puncte de întâlnire în caz de cutremur.";
+"CreateGroup.firstParagraph" = "În eventualitatea unui cutremur este important să ai un protocol împreună cu cei apropiați prin care vă puteți regăsi cu ușurință. Creează un grup de prieteni, de familie sau de colegi de la birou sau școală pentru a vă seta puncte de întâlnire în caz de cutremur.";
 
 "CreateGroup.secondParagraph" = "Chiar dacă nu vei avea acces la internet în acele momente, poți folosi în continuare aplicație pentru a vedea care sunt punctele de întâlnire și pentru a identifica traseele pentru a ajunge acolo. De asemenea, poți vedea, cu permisiunea lor, traseele și adresele uzuale ale membrilor grupului tău și îi poți căuta dacă aceștia nu vin la punctul de întâlnire.";
